refactor: 모달 사용성 개선
