+++ app/(protected)/(tabs)/history.tsx
@@ -10,6 +10,7 @@ import {
 
 import CalendarNavigator from "@/components/CalendarNavigator";
 import WorkoutCalendar from "@/components/WorkoutCalendar";
+import RestDayModal from "@/components/modals/RestDayModal";
 import icons from "@/constants/icons";
 import useCalendar from "@/hooks/useCalendar";
 import useFetchData from "@/hooks/useFetchData";
@@ -79,7 +80,7 @@ export default function History() {
 
         <SetRestDayButton
           onPress={() => {
-            openModal({ type: "REST_DAY" }, "bottom");
+            openModal(<RestDayModal />, "bottom");
           }}
         />
       </View>
